Misc Tests, Fixes, and Features (#50)
* Stop event popagation and default behavior

To accomplish this we need to propagation and default behavior
flags to the event spec and add logic to JS in order to handle them.

Adds a test for this via drag-and-drop with selenium. Not working
now, but should be able to get it working.

* Fix bug in animation where you could register a new animation hook
after stopping the original hook (i.e. by returning False). We
solve this by marking whether or not an element's animation hook
has stopped. If you register a hook after a prior one has stopped
then a RuntimeError will be raised.

We also change the animation API by passing in a stop function.
Once the stop function has been called then the animation hook
will not be called again and no new animation hooks can be
registered to the element.

* Remove the drag and drop test

Drag and drop not working for some reason via Selenium. It works
when done manually, but selenium doesn't seem able to move the
cursor after a drag has been initiated.

* you can configure event without callback