no need for raw coro to set/unset current hook
since element updates are processed sequentially and not
in parallel we don't need the dynamic setting/unsetting
or the current hook. in the future we may need to return
to that approach, but for now sequential updates are
simpler, more performant when processing individual updates
(since no logic between each yield), and maybe more aligned
with how React works.