federate modules with mount function
Allows modules to define a mount function that can attach
components directly to a DOM element. This gives authors of
custom components a high degree of flexibility and alleviates
the problem have having conflicting React instances when
idom-client-react is installed independently.

For example, custom components with a mount function need
not be implemented in React.