fix #420 warn for 'key' param in render function
previously key was a required parameter if it was given. initially
we thought that just making it optional would be fine. however,
after some thought, this could lead to unexpected behavior when
users don't want 'key' to indicate identity - if it were optional
the user would not get any warning when this confusion occured. As
a result we will warn the user about this until, in a future
release we raise an exception.